# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Client Portal API for freelance design business
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects (admin sees all, clients see their own)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, in_progress, review, completed, cancelled]
        - name: clientId
          in: query
          schema:
            type: ["integer", "null"]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectListResponse"

    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a single project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Project Requests (client-submitted requests)
  /requests:
    get:
      operationId: listRequests
      tags: [requests]
      summary: List project requests
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestListResponse"

    post:
      operationId: createRequest
      tags: [requests]
      summary: Submit a new project request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRequestBody"
      responses:
        "201":
          description: Created request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectRequest"
  /requests/{id}:
    get:
      operationId: getRequest
      tags: [requests]
      summary: Get a single request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectRequest"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateRequest
      tags: [requests]
      summary: Update or approve/reject a request (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRequestBody"
      responses:
        "200":
          description: Updated request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectRequest"

  # Files
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Project:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [pending, in_progress, review, completed, cancelled]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        clientId:
          type: integer
        clientName:
          type: string
        budget:
          type: ["number", "null"]
        deadline:
          type: ["string", "null"]
          format: date
        completedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - status
        - priority
        - clientId
        - clientName
        - createdAt
        - updatedAt
    ProjectDetail:
      allOf:
        - $ref: "#/components/schemas/Project"
        - type: object
          properties:
            files:
              type: array
              items:
                $ref: "#/components/schemas/ProjectFile"
            comments:
              type: array
              items:
                $ref: "#/components/schemas/Comment"
          required:
            - files
            - comments
    ProjectListResponse:
      type: object
      properties:
        projects:
          type: array
          items:
            $ref: "#/components/schemas/Project"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - projects
        - total
        - page
        - limit
    CreateProjectBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        clientId:
          type: integer
        status:
          type: string
          enum: [pending, in_progress, review, completed, cancelled]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        budget:
          type: ["number", "null"]
        deadline:
          type: ["string", "null"]
          format: date
      required:
        - title
        - description
        - clientId
    UpdateProjectBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [pending, in_progress, review, completed, cancelled]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        budget:
          type: ["number", "null"]
        deadline:
          type: ["string", "null"]
          format: date
    ProjectRequest:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        serviceType:
          type: string
          enum: [logo_design, branding, web_design, print_design, social_media, illustration, other]
        budget:
          type: ["number", "null"]
        timeline:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        clientId:
          type: integer
        clientName:
          type: string
        adminNotes:
          type: ["string", "null"]
        projectId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - serviceType
        - status
        - clientId
        - clientName
        - createdAt
        - updatedAt
    RequestListResponse:
      type: object
      properties:
        requests:
          type: array
          items:
            $ref: "#/components/schemas/ProjectRequest"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - requests
        - total
        - page
        - limit
    CreateRequestBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        serviceType:
          type: string
          enum: [logo_design, branding, web_design, print_design, social_media, illustration, other]
        budget:
          type: ["number", "null"]
        timeline:
          type: string
      required:
        - title
        - description
        - serviceType
    UpdateRequestBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, approved, rejected]
        adminNotes:
          type: string
        projectId:
          type: ["integer", "null"]
    ProjectFile:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        name:
          type: string
        url:
          type: string
        size:
          type: ["integer", "null"]
        mimeType:
          type: ["string", "null"]
        uploadedBy:
          type: integer
        uploaderName:
          type: string
        uploaderRole:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - name
        - url
        - uploadedBy
        - uploaderName
        - uploaderRole
        - createdAt
...
    Comment:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        content:
          type: string
        authorId:
          type: integer
        authorName:
          type: string
        authorRole:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - content
        - authorId
        - authorName
        - authorRole
        - createdAt
...