# lib/api-zod/src/generated/types/updateProjectBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Client Portal API for freelance design business
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateProjectBodyPriority } from "./updateProjectBodyPriority";
import type { UpdateProjectBodyStatus } from "./updateProjectBodyStatus";

export interface UpdateProjectBody {
  title?: string;
  description?: string;
  status?: UpdateProjectBodyStatus;
  priority?: UpdateProjectBodyPriority;
  /** @nullable */
  budget?: number | null;
  /** @nullable */
  deadline?: Date | null;
}